+++ system/modules/isotope/IsotopeOrder.php
@@ -527,7 +527,7 @@ class IsotopeOrder extends IsotopeProductCollection
 
 		if ($this->strOrderId == '')
 		{
-			$strPrefix = $this->Isotope->Config->orderPrefix;
+			$strPrefix = $this->Isotope->replaceInsertTags($this->Isotope->Config->orderPrefix);
 			$intPrefix = utf8_strlen($strPrefix);
 			$arrConfigIds = $this->Database->execute("SELECT id FROM tl_iso_config WHERE store_id=" . $this->Isotope->Config->store_id)->fetchEach('id');