+++ system/modules/isotope/dca/tl_iso_products.php
@@ -48,7 +48,6 @@ $GLOBALS['TL_DCA']['tl_iso_products'] = array
 		(
 			array('tl_iso_products', 'applyAdvancedFilters'),
 			array('tl_iso_products', 'checkPermission'),
-			array('tl_iso_products', 'addBreadcrumb'),
 			array('tl_iso_products', 'buildPaletteString'),
 		),
 		'onsubmit_callback' => array
@@ -751,116 +750,6 @@ class tl_iso_products extends Backend
 			$this->log('Cannot access product ID '.$this->Input->get('id'), __METHOD__, TL_ACCESS);
 			$this->redirect('contao/main.php?act=error');
 		}
-	}
-
-
-	/**
-	 * Add the breadcrumb menu
-	 * @return void
-	 */
-	public function addBreadcrumb()
-	{
-		// Set a new node
-		if (isset($_GET['node']))
-		{
-			$this->Session->set('tl_page_node', $this->Input->get('node'));
-			$this->redirect(preg_replace('/&node=[^&]*/', '', $this->Environment->request));
-		}
-
-		$intNode = $this->Session->get('tl_page_node');
-
-		if ($intNode < 1)
-		{
-			return;
-		}
-
-		$arrIds = array();
-		$arrLinks = array();
-
-		// Generate breadcrumb trail
-		if ($intNode)
-		{
-			$this->loadDataContainer('tl_page');
-			$tl_page = new tl_page();
-			$intId = $intNode;
-
-			do
-			{
-				$objPage = $this->Database->prepare("SELECT * FROM tl_page WHERE id=?")
-								->limit(1)
-								->execute($intId);
-
-				if ($objPage->numRows < 1)
-				{
-					// Currently selected page does not exits
-					if ($intId == $intNode)
-					{
-						$this->Session->set('tl_page_node', 0);
-						return;
-					}
-
-					break;
-				}
-
-				$arrIds[] = $intId;
-
-				// No link for the active page
-				if ($objPage->id == $intNode)
-				{
-					$arrLinks[] = $tl_page->addIcon($objPage->row(), '', null, '', true) . ' ' . $objPage->title;
-				}
-				else
-				{
-					$arrLinks[] = $tl_page->addIcon($objPage->row(), '', null, '', true) . ' <a href="' . $this->addToUrl('node='.$objPage->id) . '">' . $objPage->title . '</a>';
-				}
-
-				// Do not show the mounted pages
-				if (!$this->User->isAdmin && in_array($objPage->id, $this->User->pagemounts))
-				{
-					break;
-				}
-
-				$intId = $objPage->pid;
-			}
-			while ($intId > 0 && $objPage->type != 'root');
-		}
-
-		// Check whether the node is mounted
-		if (!$this->User->isAdmin && !$this->User->hasAccess($arrIds, 'pagemounts'))
-		{
-			$this->Session->set('tl_page_node', 0);
-
-			$this->log('Page ID '.$intNode.' was not mounted', 'tl_page addBreadcrumb', TL_ERROR);
-			$this->redirect($this->Environment->script.'?act=error');
-		}
-
-		// Limit tree
-		$arrNodes = array_merge(array($intNode), $this->getChildRecords($intNode, 'tl_page'));
-		$objProducts = $this->Database->execute("SELECT pid FROM tl_iso_product_categories WHERE page_id IN (" . implode(',', $arrNodes) . ")");
-
-		if ($objProducts->numRows)
-		{
-			$GLOBALS['TL_DCA']['tl_iso_products']['list']['sorting']['root'] = $objProducts->fetchEach('pid');
-		}
-		else
-		{
-			$berror = '
-
-<ul id="tl_breadcrumb">
-  <li>' . $GLOBALS['TL_LANG']['ERR']['breadcrumbEmpty'] . '</li>
-</ul>';
-		}
-
-		// Add root link
-		$arrLinks[] = '<img src="system/themes/' . $this->getTheme() . '/images/pagemounts.gif" width="18" height="18" alt=""> <a href="' . $this->addToUrl('node=0') . '">' . $GLOBALS['TL_LANG']['MSC']['filterAll'] . '</a>';
-		$arrLinks = array_reverse($arrLinks);
-
-		// Insert breadcrumb menu
-		$GLOBALS['TL_DCA']['tl_iso_products']['list']['sorting']['breadcrumb'] = '
-
-<ul id="tl_breadcrumb">
-  <li>' . implode(' &gt; </li><li>', $arrLinks) . '</li>
-</ul>'.$berror;
 	}