+++ system/modules/isotope/CHANGELOG.md
@@ -17,6 +17,7 @@ Version 1.4.beta1 (2012-??-??)
 - Product list can be sorted by dateAdded, this allows to list the newest products
 - More modular gallery class, also allows to retrieve the list of images (#3173)
 - Added getimagesize data to gallery template (#3173)
+- Added edit-header dropdown for mail templates (#3228)
 
 
 Version 1.3.2 stable (2012-06-11)