+++ system/modules/isotope/CHANGELOG.md
@@ -13,6 +13,7 @@ Version 1.4.beta1 (2012-??-??)
 - Support for i18nl10n extension
 - The backend now lists new orders on the welcome screen (#2111)
 - Ever product can now have a CSS class and ID (#2812)
+- Rules can now be limited to subtotal in cart
 
 ### Improved
 - Checkout steps are now a numbered list (#3040)